% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaddress.R
\name{gaddress}
\alias{gaddress}
\title{gaddress}
\usage{
gaddress(address)
}
\arguments{
\item{address}{Character class of location address}
}
\value{
A data.frame contains address, latitude and longitude of given location.
}
\description{
Returns latitude and longitude of the given input location.
}
\examples{
gaddress("37.56654, 126.978")
gaddress("37.55723, 127.0453")
}
