\name{ggeor-package}
\alias{ggeor-package}
\alias{ggeor}
\docType{package}
\title{
\packageTitle{ggeor}
}
\description{
\packageDescription{ggeor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ggeor}
\packageIndices{ggeor}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ggeor}

Maintainer: \packageMaintainer{ggeor}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
