% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatlng.R
\name{glatlng}
\alias{glatlng}
\title{glatlng}
\usage{
glatlng(latlng)
}
\arguments{
\item{latlng}{Character class of latitude and longitude separated by ,}
}
\value{
Character class of the address of the given latitude and longitude.
}
\description{
Returns address of the given latitude and longitude.
}
\examples{
glatlng("37.56654, 126.978")
glatlng("37.55723, 127.0453")
}
