\name{ggeor-package}
\alias{ggeor-package}
\alias{ggeor}
\docType{package}
\title{
ggeor package
}
\description{
\packageDescription{ggeor}
}

\author{
\packageAuthor{ggeor}

Maintainer: \packageMaintainer{ggeor}
}
\references{
\url{https://developers.google.com/maps/documentation/geocoding/intro}
}
\keyword{ package }
\examples{
gaddress("Seoul")
gaddress("Hanyang University")
glatlng("37.56654, 126.978")
glatlng("37.55723, 127.0453")
}
